Fix setting kernel_name when None
I found out (and I didn't test this before) that ExecutePreprocessor
uses Traitlets to validate arguments at runtime, and setting kernel_name
to None to get the default doesn't work; we need to simply not set
kernel_name. So this is the most pragmatic solution: just having two
different calls to ExecutePreprocessor. If we have more arguments we
could switch to building an arguments dict.